guard ~/.config from being clobbered

+++ opencode/opencode.rake
@@ -2,16 +2,20 @@ require 'fileutils'
 
 namespace :opencode do
   desc 'Install custom opencode files'
-  task :install do
+  task :install => :ensure_dirs do
     parent_repo     = File.expand_path('~/repos/github.com/copiousfreetime')
     repo_dir        = File.join(parent_repo, 'opencode-files')
     config_dir      = File.expand_path('~/.config')
     opencode_config = File.join(config_dir, 'opencode')
 
-    FileUtils.mkdir_p(parent_repo)
-    FileUtils.mkdir_p(config_dir)
+    %x[ git clone git@github.com:copiousfreetime/opencode-files.git #{repo_dir} ] unless File.directory?(repo_dir)
 
-    %x[ git clone git@github.com:copiousfreetime/opencode-files.git #{repo_dir} ]
-    %x[ ln -s #{repo_dir} #{opencode_config} ]
+    if File.exist?(opencode_config) then
+      if !File.symlink?(opencode_config) then
+        puts "#{opencode_config} is not a symlink, this is a problem"
+      end
+    else
+      %x[ ln -s #{repo_dir} #{opencode_config} ]
+    end
   end
 end
