+++ vim/vim.rake
@@ -3,17 +3,21 @@ require 'open-uri'
 
 namespace :vim do
   desc "Install custom vim actions"
-  task :install do
+  task :install => :ensure_dirs do
     parent_repo = File.expand_path("~/repos/github.com/copiousfreetime")
     repo_dir    = File.join(parent_repo, "vimfiles")
     config_dir  = File.expand_path("~/.config")
     nvim_config = File.join(config_dir, "nvim")
 
-    FileUtils.mkdir_p(parent_repo)
-    FileUtils.mkdir_p(config_dir)
+    %x[ git clone git@github.com:copiousfreetime/vimfiles.git #{repo_dir} ] unless File.directory?(repo_dir)
 
-    %x[ git clone git@github.com:copiousfreetime/vimfiles.git #{repo_dir} ]
-    %x[ ln -s #{repo_dir} #{nvim_config} ]
+    if File.exist?(nvim_config) then
+      if !File.symlink?(nvim_config) then
+        puts "#{nvim_config} is not a symlink, this is a problem"
+      end
+    else
+      %x[ ln -s #{repo_dir} #{nvim_config} ]
+    end
   end
 
   desc "Install vim itself"
